remove unuse component
