+++ src/pages/Pool-V3/components/CreateNewPosition/index.tsx
@@ -2,20 +2,18 @@ import { PoolWithPoolKey } from '@oraichain/oraidex-contracts-sdk/build/Oraiswap
 import { poolKeyToString } from '@oraichain/oraiswap-v3';
 import { ReactComponent as CloseIcon } from 'assets/icons/close.svg';
 import classNames from 'classnames';
-import TooltipHover from 'components/TooltipHover';
 import { oraichainTokens } from '@oraichain/oraidex-common';
 import { getIcon } from 'helper';
 import useTheme from 'hooks/useTheme';
 import { useRef, useState } from 'react';
-import CreatePositionFormNew from '../CreatePositionFormNew';
 import styles from './index.module.scss';
 import cn from 'classnames/bind';
-import { reduceString } from 'libs/utils';
 import { useGetPositions } from 'pages/Pool-V3/hooks/useGetPosition';
 import useConfigReducer from 'hooks/useConfigReducer';
 import { extractAddress } from 'pages/Pool-V3/helpers/format';
 import { useDispatch } from 'react-redux';
 import { setToDefault } from 'reducer/poolDetailV3';
+import CreatePositionForm from '../CreatePositionForm';
 
 const cx = cn.bind(styles);
 export const openInNewTab = (url: string): void => {
@@ -128,7 +126,7 @@ const CreateNewPosition = ({
               </div>
             </div>
           </div>
-          <CreatePositionFormNew
+          <CreatePositionForm
             poolId={poolKeyToString(pool.pool_key)}
             showModal={showModal}
             slippage={slippage}
